Add syntactic sugar
