+++ CHANGELOG.md
@@ -25,4 +25,3 @@ All notable changes to this project will be documented in this file based on [Ke
 ***Added 🌿***
 
 - Initial release
-
