+++ NEWS.d/20250221_191040_makukha_9_overload_class_lshift_as_pretty_constructor.md
@@ -1,0 +1,6 @@
+# ***Added 🌿***
+
+- New methods to append entries: `__or__()`, `__ior__()`, `add()`, `update()`.
+- New method `copy()` to get a deep copy.
+- Method `__contains__()` to check whether directory layout object contains specific path.
+- Item access with `__getitem__()` returning `Path` object.
