+++ docs/features.md
@@ -3,4 +3,5 @@
 - Create and remove directory tree with files
 - Chdir to layout subdirectories
 - Display as rich tree for docs
+- Syntactic sugar: mapping interface, context manager, append with `|` and `|=`
 - Uses [pathlib2](https://github.com/jazzband/pathlib2) for Python 2
