+++ src/dirlay/__init__.pyi
@@ -1,8 +1,9 @@
-from pathlib import Path
+import pathlib
 from typing import Any, Iterable, Iterator, Optional, Tuple, Union
 
 from typing_extensions import TypeAlias
 
+Path: TypeAlias = pathlib.Path
 PathType: TypeAlias = Union[Path, str]
 
 NodeTree: TypeAlias = dict[PathType, 'NodeValue']
@@ -16,9 +17,18 @@ class DirLayout:
     _basedir: Optional[Path]
     _prevdir: Optional[str]
     def __init__(self, entries: Optional[NodeTree] = ...) -> None: ...
+    def __contains__(self, path: PathType) -> bool: ...
     def __eq__(self, other: Any) -> bool: ...
+    def __getitem__(self, path: str) -> Path: ...
     def __iter__(self) -> Iterator[Tuple[Path, Optional[str]]]: ...
+    def __or__(self, entries: NodeTree) -> 'DirLayout': ...
+    def __ior__(self, entries: NodeTree) -> 'DirLayout': ...
+    def __enter__(self) -> 'DirLayout': ...
+    def __exit__(self, exc_type: Any, exc_val: Any, exc_tb: Any) -> 'DirLayout': ...
+    def add(self, entries: NodeTree) -> 'DirLayout': ...
+    def copy(self) -> 'DirLayout': ...
     def to_dict(self) -> StrNodeTree: ...
+    def update(self, entries: NodeTree) -> None: ...
     @classmethod
     def _add_path(
         cls,
